requirements.txt
python-telegram-bot==20.7
mercadopago==2.2.0
APScheduler==3.10.4
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0